make stage for storage starbursts in content view

+++ src/components/content/content-item.jsx
@@ -2,10 +2,14 @@ import React from 'react'
 
 import ContentItemList from './content-item-list'
 import ContentItemChart from './content-item-chart'
+import StorageChartContainer from '../../containers/storage/storage-chart-container'
 import StorageItemList from './storage/storage-item-list'
 
 const ContentItem = ({ deleteItem, isChart, isStorage, itemProps }) => {
   if (isChart) {
+    if (isStorage) {
+      return <StorageChartContainer id={itemProps.id} params={itemProps.params} />
+    }
     return (
       <ContentItemChart {...itemProps} delete={deleteItem} />
     )
